crowd-edge-cloud/functions/processor/requirements.txt
functions-framework==3.4.0
google-cloud-firestore==2.15.0
google-cloud-pubsub==2.21.0
